+++ postcss.config.cjs
@@ -1,0 +1,15 @@
+const prefixSelector = require('postcss-prefix-selector');
+
+module.exports = {
+    plugins: [
+        prefixSelector({
+            prefix: '.osd-paperjs-annotation',
+            transform(prefix, selector, prefixedSelector) {
+                if (selector === '.annotation-ui-noselect') {
+                    return 'body.annotation-ui-noselect';
+                }
+                return prefixedSelector;
+            },
+        }),
+    ],
+};