+++ webpack.config.js
@@ -66,7 +66,16 @@ module.exports = {
             },
             {
                 test: /\.css$/,
-                use: ['style-loader', 'css-loader']
+                use: [
+                    {
+                        loader: 'style-loader',
+                        options: {
+                            attributes: { 'data-osd-paperjs-annotation': 'true' },
+                        },
+                    },
+                    'css-loader',
+                    'postcss-loader',
+                ],
             },
             {
                 test: /\.png$/,